Consolidate the hermes e2e tests
